refactor: extract NoteCell and NoteSideMenu common code to helper hooks
